+++ src/main/scala/net/exoego/facade/aws_lambda/alb/package.scala
@@ -1,0 +1,6 @@
+package net.exoego.facade.aws_lambda
+
+package object alb {
+  type ALBHandler = Handler[ALBEvent, ALBResult]
+  type AsyncALBHandler = AsyncHandler[ALBEvent, ALBResult]
+}
