+++ src/main/scala/net/exoego/facade/aws_lambda/apigateway/proxy/package.scala
@@ -1,0 +1,10 @@
+package net.exoego.facade.aws_lambda.apigateway
+
+import net.exoego.facade.aws_lambda.{AsyncHandler, Handler}
+
+package object proxy {
+  type APIGatewayProxyHandler =
+    Handler[APIGatewayProxyEvent, APIGatewayProxyResult]
+  type AsyncAPIGatewayProxyHandler =
+    AsyncHandler[APIGatewayProxyEvent, APIGatewayProxyResult]
+}
