+++ src/main/scala/net/exoego/facade/aws_lambda/cloudfront/package.scala
@@ -1,0 +1,7 @@
+package net.exoego.facade.aws_lambda
+
+import scala.scalajs.js
+
+package object cloudfront {
+  type CloudFrontHeaders = js.Dictionary[js.Array[CloudFrontHeadersItem]]
+}
