+++ src/main/scala/net/exoego/facade/aws_lambda/codepipeline/cloudwatch/pipeline/package.scala
@@ -1,0 +1,11 @@
+package net.exoego.facade.aws_lambda.codepipeline.cloudwatch
+
+import net.exoego.facade.aws_lambda.{AsyncHandler, Handler}
+
+package object pipeline {
+  type CodePipelineState = String
+  type CodePipelineCloudWatchPipelineHandler =
+    Handler[CodePipelineCloudWatchPipelineEvent, Unit]
+  type AsyncCodePipelineCloudWatchPipelineHandler =
+    AsyncHandler[CodePipelineCloudWatchPipelineEvent, Unit]
+}
