+++ src/main/scala/net/exoego/facade/aws_lambda/codepipeline/cloudwatch/stage/package.scala
@@ -1,0 +1,11 @@
+package net.exoego.facade.aws_lambda.codepipeline.cloudwatch
+
+import net.exoego.facade.aws_lambda.{AsyncHandler, Handler}
+
+package object stage {
+  type CodePipelineStageState = String
+  type CodePipelineCloudWatchStageHandler =
+    Handler[CodePipelineCloudWatchStageEvent, Unit]
+  type AsyncCodePipelineCloudWatchStageHandler =
+    AsyncHandler[CodePipelineCloudWatchStageEvent, Unit]
+}
