+++ src/main/scala/net/exoego/facade/aws_lambda/codepipeline/package.scala
@@ -1,0 +1,7 @@
+package net.exoego.facade.aws_lambda
+
+package object codepipeline {
+  type CodePipelineHandler = Handler[CodePipelineEvent, Unit]
+  type AsyncCodePipelineHandler = AsyncHandler[CodePipelineEvent, Unit]
+  type ArtifactLocation = S3ArtifactStore
+}
