+++ src/main/scala/net/exoego/facade/aws_lambda/dynamodbstream/package.scala
@@ -1,0 +1,7 @@
+package net.exoego.facade.aws_lambda
+
+package object dynamodbstream {
+  type AttributeValue = facade.amazonaws.services.dynamodb.AttributeValue
+  type DynamoDBStreamHandler = Handler[DynamoDBStreamEvent, Unit]
+  type AsyncDynamoDBStreamHandler = AsyncHandler[DynamoDBStreamEvent, Unit]
+}
