+++ src/main/scala/net/exoego/facade/aws_lambda/firehose/package.scala
@@ -1,0 +1,10 @@
+package net.exoego.facade.aws_lambda
+
+package object firehose {
+  type FirehoseRecordTransformationStatus = String
+  type FirehoseTransformationHandler =
+    Handler[FirehoseTransformationEvent, FirehoseTransformationResult]
+  type AsyncFirehoseTransformationHandler =
+    AsyncHandler[FirehoseTransformationEvent, FirehoseTransformationResult]
+
+}
