+++ src/main/scala/net/exoego/facade/aws_lambda/kinesisstream/package.scala
@@ -1,0 +1,6 @@
+package net.exoego.facade.aws_lambda
+
+package object kinesisstream {
+  type KinesisStreamHandler = Handler[KinesisStreamEvent, Unit]
+  type AsyncKinesisStreamHandler = AsyncHandler[KinesisStreamEvent, Unit]
+}
