+++ src/main/scala/net/exoego/facade/aws_lambda/lex/package.scala
@@ -1,0 +1,10 @@
+package net.exoego.facade.aws_lambda
+
+import scala.scalajs.js.|
+
+package object lex {
+  type LexDialogAction =
+    LexDialogActionClose | LexDialogActionElicitIntent | LexDialogActionElicitSlot | LexDialogActionConfirmIntent | LexDialogActionDelegate
+  type LexHandler = Handler[LexEvent, LexResult]
+  type AsyncLexHandler = AsyncHandler[LexEvent, LexResult]
+}
