+++ src/main/scala/net/exoego/facade/aws_lambda/s3batch/package.scala
@@ -1,0 +1,7 @@
+package net.exoego.facade.aws_lambda
+
+package object s3batch {
+  type S3BatchHandler = Handler[S3BatchEvent, S3BatchResult]
+  type AsyncS3BatchHandler = AsyncHandler[S3BatchEvent, S3BatchResult]
+  type S3BatchResultResultCode = String
+}
