+++ src/main/scala/net/exoego/facade/aws_lambda/sns/package.scala
@@ -1,0 +1,10 @@
+package net.exoego.facade.aws_lambda
+
+import scala.scalajs.js
+
+package object sns {
+  type SNSHandler = Handler[SNSEvent, Unit]
+  type AsyncSNSHandler = AsyncHandler[SNSEvent, Unit]
+  type SNSMessageAttributes = js.Dictionary[SNSMessageAttribute]
+
+}
